+++ CHANGELOG.md
@@ -13,6 +13,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 
+## [0.7.0]
+
+### Added
+  - PODAAC-5424
+    - Update concise test in uat to test only POCLOUD collections
+  - Publishing L2ss Concise Chain umm-s Update
+### Changed 
+  - Issue-68
+    - Updated jupyter notebook
+    - Update notebook test to use python code directly instead of using jupyter notebook
+  - Updated python libraries
+  - Update history json to have url in history
+### Deprecated 
+### Removed
+### Fixed
+
+
 ## [0.6.1]
 
 ### Added